backend/app/requirements.txt
Flask==3.0.3
flask-restx==1.3.0
Werkzeug==3.0.3
minio==7.1.13
flask-cors==4.0.0
python-dotenv==1.0.0
